requirements.txt
django-cc
